Adapt noise generation to add more unknownwords
